+++ confmap/provider/envprovider/provider.go
@@ -20,9 +20,8 @@ import (
 	"os"
 	"strings"
 
-	"gopkg.in/yaml.v3"
-
 	"go.opentelemetry.io/collector/confmap"
+	"go.opentelemetry.io/collector/confmap/provider/internal"
 )
 
 const schemeName = "env"
@@ -39,16 +38,9 @@ func New() confmap.Provider {
 
 func (emp *provider) Retrieve(_ context.Context, uri string, _ confmap.WatcherFunc) (confmap.Retrieved, error) {
 	if !strings.HasPrefix(uri, schemeName+":") {
-		return confmap.Retrieved{}, fmt.Errorf("%v uri is not supported by %v provider", uri, schemeName)
+		return confmap.Retrieved{}, fmt.Errorf("%q uri is not supported by %q provider", uri, schemeName)
 	}
-
-	content := os.Getenv(uri[len(schemeName)+1:])
-	var data map[string]interface{}
-	if err := yaml.Unmarshal([]byte(content), &data); err != nil {
-		return confmap.Retrieved{}, fmt.Errorf("unable to parse yaml: %w", err)
-	}
-
-	return confmap.NewRetrievedFromMap(confmap.NewFromStringMap(data)), nil
+	return internal.NewRetrievedFromYAML([]byte(os.Getenv(uri[len(schemeName)+1:])))
 }
 
 func (*provider) Scheme() string {